migrate Button tests
